Rename test_plugin package
- Rename testing plugin package from `com.adjust.sdktesting` to
`com.adjust.testingmodule`
- Remove placeholder `index.js` in test_plugin
- Fix typos
