+++ test_plugin/android/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.adjust.testingmodule">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+
+</manifest>
