+++ test_plugin/package.json
@@ -2,15 +2,14 @@
     "name": "react-native-adjust-testing",
     "description": "Adjust Testing React Native SDK",
     "version": "1.0.0",
-    "main": "index.js",
     "author": "adjust GmbH",
     "license": "MIT",
     "keywords": [
         "react-native",
         "ios",
         "android",
-        "adjust"
-        "testing",
+        "adjust",
+        "testing"
     ],
     "repository": {
         "type": "git",
